portfolio_website/projects/global_weather_tracker/requirements.txt
flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.0
werkzeug==2.3.6
flask-cors==4.0.0
